update get user by team , get todoItem by team, put team, delete team

+++ aspApi/Controllers/TeamsController.cs
@@ -28,67 +28,92 @@ namespace aspApi.Controllers
         }
 
         // GET: api/Teams
-        [HttpGet]
+       /* [HttpGet]
         [Authorize(Roles = "Admin, User")]
 
         public async Task<ActionResult<IEnumerable<Team>>> GetTeams()
         {
             return await _context.Teams.ToListAsync();
-        }
+        }*/
 
         // GET: api/Teams/5
-       /* [HttpGet("{id}")]
-        [Authorize(Roles = "Admin, User")]
-        public async Task<ActionResult<Team>> GetTeam(int id)
-        {
-            var team = await _context.Teams.FindAsync(id);
+        /* [HttpGet("{id}")]
+         [Authorize(Roles = "Admin, User")]
+         public async Task<ActionResult<Team>> GetTeam(int id)
+         {
+             var team = await _context.Teams.FindAsync(id);
 
-            if (team == null)
-            {
-                return NotFound();
-            }
+             if (team == null)
+             {
+                 return NotFound();
+             }
 
-            return team;
-        }*/
+             return team;
+         }*/
 
         // PUT: api/Teams/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> PutTeam(int id, TeamDTO teamDTO)
         {
-            if (id != teamDTO.TeamId)
-            {
-                return BadRequest("Id mismatch between URL and request body.");
-            }
+            var team = await _context.Teams
+                .Include(t => t.TeamUsers)
+                .FirstOrDefaultAsync(t => t.TeamId == id);
 
-            var team = await _context.Teams.FindAsync(id);
             if (team == null)
             {
-                return NotFound();
+                return NotFound("Team not found");
             }
 
-            // Update team properties with values from teamDTO
-            team.Name = teamDTO.Name;
-            team.IsPublic = teamDTO.IsPublic;
-
-            try
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (!userId.HasValue)
             {
-                await _context.SaveChangesAsync();
+                return NotFound("User not login");
             }
-            catch (DbUpdateConcurrencyException)
+            var teamUsers = await _context.TeamUsers.Where(t => t.TeamId == id).ToListAsync();
+            if (teamUsers == null || !teamUsers.Any())
             {
-                if (!TeamExists(id))
+                Console.WriteLine("No team users found for the team");
+            }
+            else
+            {
+
+                foreach (var teamUser in teamUsers)
                 {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
+                    if (teamUser.UserId == userId && teamUser.TeamId == id)
+                    {
+                        if (teamUser.Role == "Admin")
+                        {
+                            team.Name = teamDTO.Name;
+                            team.IsPublic = teamDTO.IsPublic;
+
+                            try
+                            {
+                                await _context.SaveChangesAsync();
+                            }
+                            catch (DbUpdateConcurrencyException)
+                            {
+                                if (!TeamExists(id))
+                                {
+                                    return NotFound();
+                                }
+                                else
+                                {
+                                    throw;
+                                }
+                            }
+                            await _context.SaveChangesAsync();
+
+                            return Ok(new { message = "Team updated successfully" });
+                        }
+                        else
+                        {
+                            return BadRequest("You arent Admin");
+                        }
+                    }
                 }
             }
-
-            return Ok(team);
+            return BadRequest("Yon dont belong to this team");
         }
 
         // POST: api/Teams
@@ -111,8 +136,6 @@ namespace aspApi.Controllers
         ////    return CreatedAtAction("GetTeam", new { id = team.TeamId }, team);
         //}
         [HttpPost]
-        [Authorize(Roles = "Admin")]
-
         public async Task<ActionResult<Team>> PostTeam(TeamDTO teamDTO)
         {
             if (!ModelState.IsValid)
@@ -133,7 +156,7 @@ namespace aspApi.Controllers
             await _context.SaveChangesAsync();
 
             // Return the created Team
-            return CreatedAtAction(nameof(GetTeams), new { id = team.TeamId }, team);
+            return team;
 
         }
 
@@ -142,36 +165,66 @@ namespace aspApi.Controllers
 
         // DELETE: api/Teams/5
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Admin")]
 
         public async Task<IActionResult> DeleteTeam(int id)
         {
-            var team = await _context.Teams.FindAsync(id);
+            var team = await _context.Teams
+     .Include(t => t.TeamUsers)
+     .FirstOrDefaultAsync(t => t.TeamId == id);
+
             if (team == null)
             {
-                return NotFound("not found TeamId");
+                return NotFound("Team not found");
             }
 
-            _context.Teams.Remove(team);
-            await _context.SaveChangesAsync();
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (!userId.HasValue)
+            {
+                return NotFound("User not login");
+            }
+            var teamUsers = await _context.TeamUsers.Where(t => t.TeamId == id).ToListAsync();
+            if (teamUsers == null || !teamUsers.Any())
+            {
+                Console.WriteLine("No team users found for the team");
+            }
+            else
+            {
 
-            return Ok($"Da xoa thanh cong teamId: {id}");
+                foreach (var teamUser in teamUsers)
+                {
+                    if (teamUser.UserId == userId && teamUser.TeamId == id)
+                    {
+                        if (teamUser.Role == "Admin")
+                        {
+                            _context.Teams.Remove(team);
+                            await _context.SaveChangesAsync();
+
+                            return Ok($"Da xoa thanh cong teamId: {id}");
+                        }
+                        else
+                        {
+                            return BadRequest("You arent Admin");
+                        }
+                    }
+                }
+            }
+            return BadRequest("Yon dont belong to this team");
         }
         public class UserInTeamDto
         {
             public int UserId { get; set; }
             public string UserName { get; set; }
             public string Role { get; set; }
-            
+
         }
 
 
-         
+
         //get user from teamId
         [HttpGet("{id}/getUser")]
         public ActionResult<IEnumerable<UserInTeamDto>> GetUsersForTeamId(int id)
         {
-            var currentUser = HttpContext.User;
+
             var team = _context.Teams
                                 .Include(t => t.TeamUsers)
                                     .ThenInclude(tu => tu.User)
@@ -186,11 +239,11 @@ namespace aspApi.Controllers
             int? userId = HttpContext.Session.GetInt32("UserId");
             if (!userId.HasValue)
             {
-                return NotFound("User not found");  
+                return NotFound("User not login");
             }
             if (!team.IsPublic)
             {
-                if (!IsUserAuthorizedForTeam(currentUser, team))
+                if (!IsUserAuthorizedForTeam(team))
                 {
                     return Forbid("This is private, you dont belong to it"); // Hoặc NotFound() tùy thuộc vào logic ứng dụng của bạn
                 }
@@ -207,6 +260,44 @@ namespace aspApi.Controllers
             return Ok(userInTeam);
 
         }
+        [HttpGet("{id}/GetToDoItems")]
+      
+        public async Task<ActionResult<IEnumerable<TodoItemDto>>> GetTodoItemsForTeam(int id)
+        {
+            var team = _context.Teams
+                                .Include(t => t.TeamUsers)
+                                    .ThenInclude(tu => tu.User)
+                                .FirstOrDefault(t => t.TeamId == id);
+
+            if (team == null)
+            {
+                return NotFound("Team not found");
+            }
+
+
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (!userId.HasValue)
+            {
+                return NotFound("User not login");
+            }
+            if (!team.IsPublic)
+            {
+                if (!IsUserAuthorizedForTeam(team))
+                {
+                    return Forbid("This is private, you dont belong to it"); // Hoặc NotFound() tùy thuộc vào logic ứng dụng của bạn
+                }
+
+            }
+            var todoItems = await _context.Teams
+                 .Where(t => t.TeamId == id)
+                 .SelectMany(t => t.TodoItems)
+                 .ToListAsync();
+
+
+            return Ok(todoItems);
+
+
+        }
         [HttpPut("{id}/privacy")]
         //[Authorize(Roles = "Admin")]
         public async Task<IActionResult> UpdateTeamPrivacy(int id, bool isPublic)
@@ -220,49 +311,43 @@ namespace aspApi.Controllers
                 return NotFound("Team not found");
             }
 
-            var userId = GetUserIdFromClaims(HttpContext.User);
-            //var userId = GetUserIdFromClaims(HttpContext.User);
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (!userId.HasValue)
+            {
+                return NotFound("User not login");
+            }
             var teamUsers = await _context.TeamUsers.Where(t => t.TeamId == id).ToListAsync();
-
             if (teamUsers == null || !teamUsers.Any())
             {
                 Console.WriteLine("No team users found for the team");
             }
             else
             {
-                
+
                 foreach (var teamUser in teamUsers)
                 {
-                    if(teamUser.UserId == userId && teamUser.TeamId ==id) {
-                        if(teamUser.Role != "Admin")
+                    if (teamUser.UserId == userId && teamUser.TeamId == id)
+                    {
+                        if (teamUser.Role == "Admin")
                         {
                             //Console.WriteLine("ko dc");
-                            return BadRequest("You don't have Permission");
+                            team.IsPublic = isPublic;
+                            await _context.SaveChangesAsync();
+
+                            return Ok(new { message = "Team privacy updated successfully" });
+                        }
+                        else
+                        {
+                            return BadRequest("You arent Admin");
                         }
                     }
                 }
             }
-
-
-            //Console.WriteLine(userId);
-
-            if(IsUserAuthorizedForTeam(HttpContext.User, team) == false)
-            {
-                return BadRequest("it doesnt belong to this team");
-            }
-
-            
-
-            team.IsPublic = isPublic;
-            await _context.SaveChangesAsync();
-
-            return Ok(new { message = "Team privacy updated successfully" });
+                    return BadRequest("Yon dont belong to this team");
         }
-
-
-        private bool IsUserAuthorizedForTeam(ClaimsPrincipal user, Team team)
+        private bool IsUserAuthorizedForTeam(Team team)
         {
-           
+
             var userId = HttpContext.Session.GetInt32("UserId").Value;
             var teamId = team.TeamId;
             Console.WriteLine(userId + " " + teamId);
@@ -271,31 +356,16 @@ namespace aspApi.Controllers
             {
                 if (i.UserId == userId && i.TeamId == teamId)
                 {
-                    Console.WriteLine(teamId + " " + userId +" " + i.Role);   
+                    Console.WriteLine(teamId + " " + userId + " " + i.Role);
                     return true;
                 }
-             
+
             }
             return false;
         }
-
-        private int GetUserIdFromClaims(ClaimsPrincipal user)
-        {
-            // Lấy UserId từ Claims của người dùng, bạn cần điều chỉnh logic này tùy thuộc vào cách bạn thực hiện xác thực
-            var userIdClaim = user.FindFirst(ClaimTypes.NameIdentifier);
-            if (userIdClaim != null && int.TryParse(userIdClaim.Value, out int userId))
-            {
-                return userId;
-            }
-            return -1; // hoặc ném ra một exception, tùy thuộc vào yêu cầu của ứng dụng
-        }
-
-
-
-
         private bool TeamExists(int id)
         {
             return _context.Teams.Any(e => e.TeamId == id);
         }
-    }
+    }   
 }
